Write to file

+++ src/Globals/functions.php
@@ -2,9 +2,14 @@
 
 use emteknetnz\TypeTransitioner\Config\Config;
 use emteknetnz\TypeTransitioner\Exceptions\TypeException;
+use SilverStripe\Core\ClassInfo;
+use SilverStripe\Core\Manifest\ClassManifest;
 
-global $_tt_method_cache;
-$_tt_method_cache = [];
+global $_ett_method_cache;
+$_ett_method_cache = [];
+
+global $_ett_lines;
+$_ett_lines = [];
 
 /**
  * Note: mixed static types are only availabe from php8.0
@@ -16,58 +21,76 @@ $_tt_method_cache = [];
 // function _a(string $castType, &$_arg)
 function _a()
 {
-    global $_tt_method_cache;
+    global $_ett_method_cache, $_ett_lines;
 
     $outdir = str_replace('//', '/', BASE_PATH . '/artifacts');
     if (!file_exists($outdir)) {
         mkdir($outdir);
     }
-    $outpath = "$outdir/tt.txt";
+    $outpath = "$outdir/ett.txt";
 
     $TYPE_DYNAMIC = 'type-dynamic'; // php7 type is blank
     $TYPE_MISSING = 'type-missing'; // dockblock return type is missing
     $WHERE_MISSING = 'where-missing'; // php7 type is blank and docblock type is missing
 
     $d = debug_backtrace(0, 2);
-    $file = $d[0]['file'] ?? '';
-    $line = $d[0]['line'] ?? '';
-    $class = $d[1]['class'] ?? '';
-    $callType = $d[1]['type'] ?? '';
-    $method = $d[1]['function'] ?? '';
-    $args = $d[1]['args'] ?? [];
-    print_r([$file, $line, $class, $callType, $method, $args]);
 
-    if (!$class || !$method) {
+    $callingFile = $d[1]['file'] ?? '';
+    $callingLine = $d[1]['line'] ?? '';
+    $calledClass = $d[1]['class'] ?? '';
+    // $callType = $d[1]['type'] ?? '';
+    $calledMethod = $d[1]['function'] ?? '';
+    $args = $d[1]['args'] ?? [];
+    // print_r([
+    //     $callingFile,
+    //     $callingLine,
+    //     // $calledLine,
+    //     $calledClass,
+    // //    $callType,
+    //     $calledMethod,
+    //     $args
+    // ]);
+
+    if (!$calledClass || !$calledMethod) {
+        return;
         echo 'No class or method';die;
     }
 
-    $key = $class . '::' . $method;
-    if (!array_key_exists($key, $_tt_method_cache)) {
-        // https://www.php.net/manual/en/book.reflection.php
-        $reflClass = new ReflectionClass($class);
-        $reflMethod = $reflClass->getMethod($method);
-        
-        $reflDocblock = $reflMethod->getDocComment();
-        $reflParams = $reflMethod->getParameters();
-        $reflReturn = $reflMethod->getReturnType();
+    $key = $calledClass . '::' . $calledMethod;
+    if (!array_key_exists($key, $_ett_method_cache)) {
 
-        preg_match_all('#@param +([^ ]+)[ \t]+((?:\$)[A-Za-z0-9]+)#', $reflDocblock, $m);
-        $docblockParams = array_combine($m[2], $m[1]);
-        $methodParams = array_combine(
-            array_map(fn($p) => '$' . $p->getName(), $reflParams),
-            array_map(fn($p) => $p->hasType() ? $p->getType()->getName() : $TYPE_DYNAMIC, $reflParams)
-        );
-        
-        preg_match('#@return ([^ ]+)#', $reflDocblock, $m);
-        $docblockReturn = $m[1] ?? $TYPE_MISSING;
-        $methodReturn = $reflReturn ? $reflReturn->getName() : '';
+        try {
 
-        $_tt_method_cache[$key] = [
-            'docblockParams' => $docblockParams,
-            'methodParams' => $methodParams,
-            'docblockReturn' => $docblockReturn,
-            'methodReturn' => $methodReturn,
-        ];
+            // https://www.php.net/manual/en/book.reflection.php
+            $reflClass = new ReflectionClass($calledClass);
+            $reflMethod = $reflClass->getMethod($calledMethod);
+            
+            $reflDocblock = $reflMethod->getDocComment();
+            $reflParams = $reflMethod->getParameters();
+            $reflReturn = $reflMethod->getReturnType();
+
+            preg_match_all('#@param +([^ ]+)[ \t]+((?:\$)[A-Za-z0-9]+)#', $reflDocblock, $m);
+            $docblockParams = array_combine($m[2], $m[1]);
+            $methodParams = array_combine(
+                array_map(fn($p) => '$' . $p->getName(), $reflParams),
+                array_map(fn($p) => $p->hasType() ? $p->getType()->getName() : $TYPE_DYNAMIC, $reflParams)
+            );
+            
+            preg_match('#@return ([^ ]+)#', $reflDocblock, $m);
+            $docblockReturn = $m[1] ?? $TYPE_MISSING;
+            $methodReturn = $reflReturn ? $reflReturn->getName() : '';
+
+            $_tt_method_cache[$key] = [
+                'namespace' => $reflClass->getNamespaceName(),
+                'docblockParams' => $docblockParams,
+                'methodParams' => $methodParams,
+                'docblockReturn' => $docblockReturn,
+                'methodReturn' => $methodReturn,
+            ];
+        } catch (Exception $e) {
+            echo "Exception when doing Refelction\n";
+            return;
+        }
     }
     $o = $_tt_method_cache[$key];
     $vars = array_keys($o['methodParams']);
@@ -75,14 +98,20 @@ function _a()
         $var = $vars[$i];
         $arg = $args[$i];
         $paramType = $o['methodParams'][$var];
+        if ($paramType != $TYPE_DYNAMIC) {
+            // strongly typed method params are a non-issue
+            continue;
+        }
         $paramWhere = 'method';
-        if ($paramType == $TYPE_DYNAMIC) {
-            if (array_key_exists($var, $o['docblockParams'])) {
-                $paramType = $o['docblockParams'][$var];
-                $paramWhere = 'docblock';
-            } else {
-                $paramWhere = $WHERE_MISSING;
+        if (array_key_exists($var, $o['docblockParams'])) {
+            $paramType = $o['docblockParams'][$var];
+            if (is_object($arg) && strpos($paramType, '\\') === false && $o['namespace']) {
+                // convert to FQCN
+                $paramType = $o['namespace'] . '\\' . $paramType;
             }
+            $paramWhere = 'docblock';
+        } else {
+            $paramWhere = $WHERE_MISSING;
         }
         if (is_null($arg)) {
             $argType = 'null';
@@ -102,11 +131,23 @@ function _a()
         if ($argType == $paramType) {
             continue;
         }
-        $line = implode(',', [$file, $line, $class, $method, $var, $paramWhere, $paramType, $argType, $arg]);
+        if ($paramWhere == 'docblock' && is_object($arg)) {
+            // is the docblock type an interface or parent class?
+            $subclasses = ClassInfo::subclassesFor($paramType);
+            echo "Subclasses are:\n";
+            var_dump($paramType);
+            print_r($subclasses);die;
+        }
+        // headers set in /app/_config.php
+        $line = implode(',', [$callingFile, $callingLine, $calledClass, $calledMethod, $var, $paramWhere, $paramType, $argType]);
+        if (array_key_exists($line, $_ett_lines)) {
+            continue;
+        }
+        $_ett_lines[$line] = true;
         file_put_contents($outpath, "$line\n", FILE_APPEND);
     }
 
-    die;
+    return;
 
     // TODO: get call stack so can provide useful exception data
     $config = Config::inst();
